backend/requirements.txt
aiomysql==0.2.0
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
atlastk==0.13.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
ecdsa==0.19.0
exceptiongroup==1.2.1
fastapi==0.110.2
h11==0.14.0
idna==3.7
iso8601==1.1.0
loguru==0.7.2
passlib==1.7.4
pyasn1==0.6.0
pydantic==2.7.1
pydantic-core==2.18.2
PyMySQL==1.1.0
pypika-tortoise==0.1.6
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tortoise-orm==0.20.0
typing-extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
